# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a pet store.",
        "title": "My Awesome API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/phrases": {
            "get": {
                "description": "Returns a list of all phrases",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phrases"
                ],
                "summary": "Get all phrases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поисковый текст",
                        "name": "text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Phrase"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new phrase to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phrases"
                ],
                "summary": "Create a new phrase",
                "parameters": [
                    {
                        "description": "New Phrase",
                        "name": "phrase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePhraseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/phrases/{id}": {
            "get": {
                "description": "Returns a phrase by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phrases"
                ],
                "summary": "Get a phrase by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Phrase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Phrase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing phrase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phrases"
                ],
                "summary": "Update a phrase by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Phrase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Phrase",
                        "name": "phrase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Phrase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Phrase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a phrase from the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "phrases"
                ],
                "summary": "Delete a phrase by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Phrase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Phrase": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "phrase_type": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                }
            }
        },
...
        "models.CreatePhraseRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "type_id": {
                    "type": "string"
                }
            }
        },
...
    }
}